+++ currency.js
@@ -1,6 +1,9 @@
 /*!
- * currency.js v0.1.3
- * github.com/scurker/currency.js
+ * currency.js - v0.1.4
+ * http://scurker.github.io/currency.js
+ *
+ * Copyright (c) 2013 Jason Wilson
+ * Released under MIT license
  */
 (function(window) {